requirements.txt
Flask==3.0.2
keras==2.10.0
matplotlib==3.8.0
numpy==1.26.4
pandas==2.2.0
Pillow==10.2.0
pymongo==4.6.1
Requests==2.31.0
tensorflow==2.10.1
opencv-python
keras_applications
bcrypt
dlib
git+https://github.com/rcmalli/keras-vggface.git